Reworked layers. Added the ability to change layers on renderers in the middle of the simulation.
This was achieved by totally rewriting the logic behind layers. Previously, when attaching a renderer to the manager, it inserted itself into a specific spot in the manageables list using binary search, so that, when iterating, the lower layers would be drawn first.
In the new version, we create a list of surfaces, and each renderer draws onto a specific surface according to its layer, those surfaces then get blitted onto the display. The surfaces are initially transparent, and get cleared every frame before anything renders.